backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0

# Authentication
google-auth==2.25.2
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
